requirements.txt
absl-py==0.8.1
astor==0.8.1
autopep8==1.4.4
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
gast==0.2.2
google-auth==1.8.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
Pillow==6.2.1
pkg-resources==0.0.0
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
python-aiml==0.9.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
